requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-rest-registration
pillow